routes for work and home pages with nav renered consistently. links to socials added.
