+++ ultimate_profile/package.json
@@ -11,7 +11,7 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "babel-plugin-macros": "^3.1.0",
-    "framer-motion": "^8.5.0",
+    "framer-motion": "^8.5.2",
     "gsap": "^3.11.4",
     "mouse-follower": "^1.1.2",
     "react": "^18.2.0",
